WIP: See is test provider can sign in
